requirements.txt
kafka-python
duckdb
python-dotenv